Add --query-string option
Sometimes you may want to add something to the query_string of each
swift request issued by getput. Say something like dealing with slo
manifests or tempurls.

This patch adds an `--query-string` option to getput, and is expected to
already be a single string ready to be apended to the request:

  --query-string "sky=blue&dog=bark"